+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Changed
+- Fix parser for multiple-words attributes: "Multiple-Words" instead of "MultipleWords"
 
 ## [0.1.0] - 2021-08-04
 ### Added